+++ 2DGameEngine/src/Components/ProjectileComponent.h
@@ -4,7 +4,7 @@
 
 struct ProjectileComponent
 {
-	ProjectileComponent(const int hitDamage = 0, const unsigned projectileDurationS = 0.0f, const bool isFriendly = false);
+	ProjectileComponent(const int hitDamage = 0, const float projectileDurationS = 0.0f, const bool isFriendly = false);
 
 	int hitDamage;
 	unsigned projectileDurationS;
